requirements.txt
django==1.9.9
celery==3.1.23
django-celery==3.1.17
pillow
sorl-thumbnail==12.4a1
